# projeto_sistema_pedidos.swagger.yaml
info:
  title: projeto_sistema_pedidos
  description: ''
  version: 1.0.0
...
paths:
  /api/clientes:
    get:
      summary: Listar Clientes
      deprecated: false
      description: ''
      operationId: get_app_cliente_listarclientes
      tags:
        - Clientes
        - Clientes
      parameters: []
      responses:
        '201':
          description: Cliente selecionado
          headers: {}
          schema:
            properties:
              id:
                type: integer
              nome:
                type: string
              email:
                type: string
            type: object
      security: []
      produces:
        - application/json
    post:
      summary: Cadastrar cliente
      deprecated: false
      description: ''
      operationId: post_app_cliente_cadastrar
      tags:
        - Clientes
        - Clientes
      parameters:
        - name: body
          in: body
          schema:
            properties:
              nome:
                required:
                  - 'true'
                type: string
              email:
                required:
                  - 'true'
                type: string
              cpf:
                required:
                  - 'true'
                type: string
            type: object
      responses:
        '201':
          description: Cliente cadastrado com sucesso
          headers: {}
          schema:
            properties:
              id:
                type: integer
              nome:
                type: string
              email:
                type: string
              cpf:
                type: string
            type: object
      security: []
      consumes:
        - application/json
      produces:
        - application/json
  /api/cliente/{cpf}:
    get:
      summary: Buscar Cliente
      deprecated: false
      description: ''
      operationId: get_app_cliente_buscarporcpf
      tags:
        - Clientes
        - Clientes
      parameters:
        - name: cpf
          in: path
          description: ''
          required: true
          type: string
          x-example: ''
      responses:
        '201':
          description: Cliente selecionado
          headers: {}
          schema:
            properties:
              id:
                type: integer
              nome:
                type: string
              email:
                type: string
            type: object
      security: []
      produces:
        - application/json
  /api/pedido:
    post:
      summary: Criar pedido
      deprecated: false
      description: ''
      operationId: post_app_pedido_criar
      tags:
        - Pedidos
        - Pedidos
      parameters:
        - name: body
          in: body
          schema:
            properties:
              cpf:
                required:
                  - 'false'
                type: string
              itens:
                type: array
                items:
                  properties:
                    produtoId:
                      type: integer
                    quantidade:
                      type: integer
                  type: object
            type: object
      responses:
        '201':
          description: Pedido criado com sucesso
          headers: {}
          schema:
            properties:
              id:
                type: integer
              status:
                type: string
              valorTotal:
                type: number
                format: float
            type: object
        '500':
          description: Falha ao fazer pedido
          headers: {}
          schema:
            properties:
              error:
                type: string
            type: object
      security: []
      consumes:
        - application/json
      produces:
        - application/json
  /api/pagamento/{pedidoId}:
    post:
      summary: Pagar pedido
      deprecated: false
      description: ''
      operationId: post_app_pedido_pagarpedido
      tags:
        - Pedidos
        - Pedidos
      parameters:
        - name: pedidoId
          in: path
          description: ''
          required: true
          type: integer
          x-example: 0
        - name: body
          in: body
          schema:
            properties:
              valor_pago:
                type: number
                format: float
                examples:
                  - 20.4
            type: object
      responses:
        '200':
          description: Status Atualizado
          headers: {}
          schema:
            properties:
              id:
                type: integer
              status:
                type: string
            type: object
        '404':
          description: Status não atualizado
          headers: {}
          schema:
            properties:
              error:
                type: string
            type: object
      security: []
      consumes:
        - application/json
      produces:
        - application/json
  /api/pedido/status/{pedidoId}:
    post:
      summary: Atualizar pedido
      deprecated: false
      description: ''
      operationId: post_app_pedido_atualizarstatus
      tags:
        - Pedidos
        - Pedidos
      parameters:
        - name: pedidoId
          in: path
          description: ''
          required: true
          type: string
          x-example: ''
        - name: body
          in: body
          schema:
            properties:
              status:
                type: string
                examples:
                  - Em Preparação
            type: object
      responses:
        '200':
          description: Status Atualizado
          headers: {}
          schema:
            properties:
              id:
                type: integer
              status:
                type: string
            type: object
        '404':
          description: Status não atualizado
          headers: {}
          schema:
            properties:
              error:
                type: string
            type: object
      security: []
      consumes:
        - application/json
      produces:
        - application/json
  /api/pedidos:
    get:
      summary: Listar pedidos
      deprecated: false
      description: ''
      operationId: get_app_pedido_listar
      tags:
        - Pedidos
        - Pedidos
      parameters: []
      responses:
        '200':
          description: Detalhes do pedido
          headers: {}
          schema:
            properties:
              id:
                type: integer
              cliente:
                type: App\Domain\Entity\Cliente
              status:
                type: string
              valor_total:
                type: number
                format: float
              criadoEm:
                type: string
              modificadoEm:
                type: string
            type: object
        '404':
          description: Pedido não encontrado
          headers: {}
          schema:
            properties:
              error:
                type: string
            type: object
      security: []
      produces:
        - application/json
  /api/acompanhar:
    get:
      summary: Acompanhar pedidos
      deprecated: false
      description: ''
      operationId: get_app_pedido_listarpedidospendentes
      tags:
        - Pedidos
        - Pedidos
      parameters: []
      responses:
        '200':
          description: Número, nome e status do pedido
          headers: {}
          schema:
            properties:
              id:
                type: integer
              nome:
                type: string
              status:
                type: string
            type: object
        '404':
          description: Pedido não encontrado
          headers: {}
          schema:
            properties:
              error:
                type: string
            type: object
      security: []
      produces:
        - application/json
...
swagger: '2.0'
...
securityDefinitions: {}
...